+++ api/src/elysia.ts
@@ -1,4 +1,11 @@
-import Elysia from "elysia";
+import { Elysia } from "elysia";
+import { entries } from "./controllers/entries";
+import { movies } from "./controllers/movies";
+import { seasonsH } from "./controllers/seasons";
+import { seed } from "./controllers/seed";
+import { series } from "./controllers/series";
+import { videosH } from "./controllers/videos";
+
 import type { KError } from "./models/error";
 
 export const base = new Elysia({ name: "base" })
@@ -30,3 +37,12 @@ export const base = new Elysia({ name: "base" })
 		return error;
 	})
 	.as("plugin");
+
+export const app = new Elysia()
+	.use(base)
+	.use(movies)
+	.use(series)
+	.use(entries)
+	.use(seasonsH)
+	.use(videosH)
+	.use(seed);
