+++ api/tests/movies/get-all-movies.test.ts
@@ -8,7 +8,7 @@ import { dune1984 } from "~/models/examples/dune-1984";
 import { dune } from "~/models/examples/dune-2021";
 import type { Movie } from "~/models/movie";
 import { isUuid } from "~/models/utils";
-import { getMovies, app } from "../helper";
+import { getMovies, app } from "../helpers";
 
 beforeAll(async () => {
 	await db.delete(shows);
