+++ api/tests/movies/get-movie.test.ts
@@ -2,7 +2,7 @@ import { beforeAll, describe, expect, it } from "bun:test";
 import { expectStatus } from "tests/utils";
 import { seedMovie } from "~/controllers/seed/movies";
 import { bubble } from "~/models/examples";
-import { getMovie } from "../helper";
+import { getMovie } from "../helpers";
 
 let bubbleId = "";
